separate and organize View files
